Implemented ARQ overflow bugž

+++ include/mpudp_worker.h
@@ -62,7 +62,7 @@ worker_t* init_worker(int, char*, monitor_t*, float);
 int spawn_worker(worker_t* spawn_worker);
 int worker_send_packet(worker_t*, mpudp_packet_t*);
 int worker_send_bcast(worker_t*, mpudp_packet_t*);
-int worker_send_ack(worker_t*, int8_t);
+int worker_send_ack(worker_t*, uint32_t);
 int watchdog_check_state(worker_t*);
 void* worker_arq_watcher(void *arg);
 #endif
